pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[project]
name = "home-assistant-mcp"
version = "0.2.0"
description = "🏠 State-of-the-Art AI-Powered Natural Language Control for Home Assistant - FastMCP 2.14.3 with Autonomous Orchestration"
readme = "README.md"
requires-python = ">=3.10,<4.0"
license = { text = "MIT" }
authors = [
    { name = "Sandra Schi", email = "sandraschipal@protonmail.com" }
]
maintainers = [
    { name = "Advanced Memory MCP Team", email = "sandraschipal@protonmail.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
    "Topic :: System :: Networking :: Monitoring",
    "Topic :: Utilities",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Operating System :: OS Independent",
    "Framework :: AsyncIO",
    "Framework :: FastAPI",
    "Framework :: Pydantic",
    "Natural Language :: English"
]
keywords = [
    "mcp",
    "model-context-protocol",
    "home-assistant",
    "smart-home",
    "automation",
    "fastmcp",
    "ai",
    "artificial-intelligence",
    "claude",
    "cursor",
    "zed",
    "natural-language",
    "conversational-ai",
    "orchestration",
    "sampling",
    "energy-optimization",
    "security",
    "iot",
    "home-automation",
    "predictive-automation",
    "multi-zone-control",
    "debugging",
    "maintenance",
    "sota",
    "state-of-the-art"
]
dependencies = [
    "fastmcp==2.14.3",              # MCP 2.14.3 stdio protocol
    "python-dotenv>=0.19.0",
    "pyyaml>=6.0",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0",
    "websockets>=12.0",
    "home-assistant-api>=0.4.0",   # HA API client
    "voluptuous>=0.13.0",          # HA config validation
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "uvicorn>=0.15.0",
    "fastapi>=0.68.0",
]

[project.urls]
Homepage = "https://github.com/sandraschi/home-assistant-mcp"
Documentation = "https://github.com/sandraschi/home-assistant-mcp#readme"
Repository = "https://github.com/sandraschi/home-assistant-mcp"
Issues = "https://github.com/sandraschi/home-assistant-mcp/issues"
Changelog = "https://github.com/sandraschi/home-assistant-mcp/blob/main/CHANGELOG.md"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "DJ", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["S101", "S104", "COM812", "ISC001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--strict-markers"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
asyncio_mode = "auto"

[project.scripts]
home-assistant-mcp = "home_assistant_mcp.cli:main"